Add 'lgian' keymap for mitosis
It's a qwerty layout, with a few tap dance key bindings. Tap dance is
used for matching brackets, e.g. () or []. The second layer has (almost)
all numbers, and symbols, along with arrow keys, and the third one has
all the miscellaneous stuff, such as F keys, volume up and down, and so
on.